requirements.txt
flask == 2.1.3
jinja2 == 3.1.2
gunicorn == 20.0.4
psycopg2 == 2.9.3
requests == 2.28.1
python-dotenv == 0.20.0
snowflake-connector-python[pandas] == 2.7.11
Flask-SQLAlchemy == 2.5.1
Flask-Migrate == 3.1.0
mysql-connector-python == 8.0.30